mylab/s04_azure_databricks/requirements.txt
azure-ai-projects
azure-ai-agents
azure-identity
databricks-ai-bridge
databricks-sdk
python-dotenv
chainlit